requirements.txt
adjustText==0.7.3
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
fire==0.4.0
fonttools==4.28.5
future==0.18.2
google-auth==2.3.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iso8601==1.0.2
itypes==1.2.0
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
many==0.6.8
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
patsy==0.5.2
Pillow==8.4.0
private==0.2.0
psycopg2==2.9.2
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
serial==0.0.97
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
statsmodels==0.12.2
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
